WIP polishing earlier solutions

+++ c/src/euler16.c
@@ -10,14 +10,14 @@
 What is the sum of the digits of the number 2^1000?
 */
 
-int digit_sum_two_power(unsigned long int power){
+unsigned digit_sum_two_power(unsigned power){
 	mpz_t result;
 	mpz_init(result);
 	mpz_ui_pow_ui(result, 2, power);
 	char* formatted = mpz_get_str(NULL, 10, result);
 	mpz_clear(result);
 	
-	int digit_sum = 0;
+	unsigned digit_sum = 0;
 	for (char* location = formatted; *location != '\0'; location++){
 		digit_sum += *location - '0';
 	}
@@ -28,6 +28,6 @@ int digit_sum_two_power(unsigned long int power){
 
 int main(){
 	assert(digit_sum_two_power(15) == 26);
-	printf("%d\n", digit_sum_two_power(1000));
+	printf("%u\n", digit_sum_two_power(1000));
 	return 0;
 }
